+++ src/components/scatter-chart/index.js
@@ -99,6 +99,7 @@ const ScatterChart = ({
       axisBottomTickValues,
       axisBottomLabelDisplayFn,
       axisLeftLabelDisplayFn,
+      typographyProps,
     ],
   )
 
@@ -125,6 +126,7 @@ const ScatterChart = ({
             { label: axisBottomLegendLabel, value: tooltipFormatX(node.data.formattedX) },
             { label: axisLeftLegendLabel, value: tooltipFormat(node.data.formattedY) },
           ] }
+          typography={ typographyProps }
         />
       ) }
       { ...getCommonProps({